+++ Assets/Scripts/HealthBar.cs
@@ -8,14 +8,14 @@ public class HealthBar : MonoBehaviour
     public Slider healthSlider;   // Reference to the UI slider
 
     // Method to set the max health value of the slider
-    public void SetMaxHealth(float health)
+    public void SetMaxHealth(int health)
     {
         healthSlider.maxValue = health;
         healthSlider.value = health;
     }
 
     // Method to update the slider with the current health value
-    public void SetHealth(float health)
+    public void SetHealth(int health)
     {
         healthSlider.value = health;
     }
